Add email notification for invoice creation in createAction function
This update integrates the Resend service to send an email notification upon invoice creation. The email template is defined in the InvoiceCreatedEmail component. Additionally, the variable name for the database insert result has been changed for clarity. The link in the email template has also been updated to point to Invoicipedia.